requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-core
langchain-google-genai
google-generativeai
faiss-cpu
PyPDF2
PyPDF
pypdf